backend/app/config.py
from pydantic import BaseModel
from dotenv import load_dotenv
import os

load_dotenv()


class Settings(BaseModel):
    app_env: str = os.getenv("APP_ENV", "development")
    port: int = int(os.getenv("PORT", "8000"))
    app_name: str = "REint Wind Forecast API"
    app_version: str = "0.1.0"


settings = Settings()