backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
httpx
python-dotenv
pyarrow
jupyter
matplotlib